refactor: DRY up stories
